feature: Move fetch(_ resource:) into new ResourceService
- Add example for fetching a named resource
